+++ src/cgdf/graphics/opengl/buffers/ebo.c
@@ -55,7 +55,7 @@ void BufferEBO_destroy(BufferEBO **ebo) {
 void BufferEBO_begin(BufferEBO *self) {
     if (!self || self->_is_begin_ || self->id == 0) return;
     glGetIntegerv(GL_ELEMENT_ARRAY_BUFFER_BINDING, &self->_id_before_begin_);
-    if (self->_id_before_begin_ != self->id) {
+    if ((uint32_t)self->_id_before_begin_ != self->id) {
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, self->id);
     }
     self->_is_begin_ = true;
@@ -64,7 +64,7 @@ void BufferEBO_begin(BufferEBO *self) {
 // Не использовать буфер:
 void BufferEBO_end(BufferEBO *self) {
     if (!self || !self->_is_begin_) return;
-    if (self->_id_before_begin_ != self->id) {
+    if ((uint32_t)self->_id_before_begin_ != self->id) {
         glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, (uint32_t)self->_id_before_begin_);
     }
     self->_is_begin_ = false;
